backend/requirements-django.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3

# Authentication
djangorestframework-simplejwt>=5.3
PyJWT>=2.8

# Database
psycopg2-binary>=2.9

# Environment
python-decouple>=3.8

# Background Tasks
celery>=5.3
redis>=5.0
django-celery-beat>=2.5
django-celery-results>=2.5

# File Storage
boto3>=1.34
django-storages>=1.14

# API & Filtering
django-filter>=23.5
drf-spectacular>=0.27

# AI & ML
openai>=1.12

# Payment Gateways
stripe>=7.12

# Utilities
python-slugify>=8.0
pillow>=10.2

# Testing
pytest>=8.0
pytest-django>=4.8

# Error Tracking (optional)
sentry-sdk>=1.40
